refactor: move automation doctypes into dedicated Automation Engine module
