+++ frappe/automation_engine/doctype/automation/automation.json
@@ -201,7 +201,7 @@
  "links": [],
  "modified": "2026-07-11 21:05:04.231189",
  "modified_by": "Administrator",
- "module": "Automation",
+ "module": "Automation Engine",
  "name": "Automation",
  "owner": "Administrator",
  "permissions": [
