+++ frappe/automation_engine/doctype/automation/automation.py
@@ -20,7 +20,7 @@ class Automation(Document):
 	from typing import TYPE_CHECKING
 
 	if TYPE_CHECKING:
-		from frappe.automation.doctype.automation_action.automation_action import AutomationAction
+		from frappe.automation_engine.doctype.automation_action.automation_action import AutomationAction
 		from frappe.types import DF
 
 		actions: DF.Table[AutomationAction]
