+++ frappe/automation_engine/doctype/automation_action/automation_action.json
@@ -38,7 +38,7 @@
  "links": [],
  "modified": "2026-07-11 21:05:04.231189",
  "modified_by": "Administrator",
- "module": "Automation",
+ "module": "Automation Engine",
  "name": "Automation Action",
  "owner": "Administrator",
  "permissions": [],
