+++ frappe/automation_engine/doctype/automation_run/automation_run.json
@@ -131,7 +131,7 @@
  "links": [],
  "modified": "2026-07-12 01:37:23.321144",
  "modified_by": "Administrator",
- "module": "Automation",
+ "module": "Automation Engine",
  "name": "Automation Run",
  "owner": "Administrator",
  "permissions": [
