+++ frappe/automation_engine/doctype/automation_run_step/automation_run_step.json
@@ -53,7 +53,7 @@
  "links": [],
  "modified": "2026-07-12 01:37:23.321144",
  "modified_by": "Administrator",
- "module": "Automation",
+ "module": "Automation Engine",
  "name": "Automation Run Step",
  "owner": "Administrator",
  "permissions": [],
