+++ frappe/automation_engine/doctype/automation_trigger_queue/automation_trigger_queue.json
@@ -117,7 +117,7 @@
  "links": [],
  "modified": "2026-07-11 20:59:51.554582",
  "modified_by": "Administrator",
- "module": "Automation",
+ "module": "Automation Engine",
  "name": "Automation Trigger Queue",
  "owner": "Administrator",
  "permissions": [
