+++ frappe/modules.txt
@@ -13,3 +13,4 @@ System
 Build
 Data
 Users
+Automation Engine
